move ocm leaves
