+++ staging/src/github.com/openshift/openshift-controller-manager/pkg/image/controller/imagestream_controller.go
