+++ staging/src/github.com/openshift/openshift-controller-manager/pkg/image/controller/metrics.go
