+++ staging/src/github.com/openshift/openshift-controller-manager/pkg/image/controller/scheduled_image_controller.go
