+++ staging/src/github.com/openshift/openshift-controller-manager/pkg/image/controller/scheduler_test.go
