+++ staging/src/github.com/openshift/openshift-controller-manager/pkg/image/controller/signature/container_image_downloader.go
