+++ staging/src/github.com/openshift/openshift-controller-manager/pkg/image/controller/signature/signature_import_controller_test.go
