+++ staging/src/github.com/openshift/openshift-controller-manager/pkg/image/controller/trigger/image_trigger_controller.go
